requirements.txt
matplotlib==3.7.0
numpy==1.24.2
Pillow==9.4.0
scikit_learn==1.2.1
torch==1.13.1
torchvision==0.14.1
